# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "VibeGo 后端服务 API",
        "title": "VibeGo API",
        "contact": {},
        "version": "0.0.1"
    },
    "host": "localhost:1984",
    "basePath": "/api",
    "paths": {
...
        "/api/git/branches": {
            "post": {
                "description": "Get list of local branches and current branch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Git"
                ],
                "summary": "List branches",
                "parameters": [
                    {
                        "description": "Path request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.GitPathRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/git/create-branch": {
            "post": {
                "description": "Create a new branch from HEAD or specified source",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Git"
                ],
                "summary": "Create branch",
                "parameters": [
                    {
                        "description": "Create branch request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateBranchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/git/delete-branch": {
            "post": {
                "description": "Delete a local branch (cannot delete current branch)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Git"
                ],
                "summary": "Delete branch",
                "parameters": [
                    {
                        "description": "Delete branch request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.DeleteBranchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/git/switch-branch": {
            "post": {
                "description": "Checkout to a different branch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Git"
                ],
                "summary": "Switch branch",
                "parameters": [
                    {
                        "description": "Switch branch request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SwitchBranchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.CreateBranchRequest": {
            "type": "object",
            "required": [
                "branch",
                "path"
            ],
            "properties": {
                "branch": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
        "handler.DeleteBranchRequest": {
            "type": "object",
            "required": [
                "branch",
                "path"
            ],
            "properties": {
                "branch": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
        "handler.GitPathRequest": {
            "type": "object",
            "required": [
                "path"
            ],
            "properties": {
                "path": {
                    "type": "string"
                }
            }
        },
...
        "handler.SwitchBranchRequest": {
            "type": "object",
            "required": [
                "branch",
                "path"
            ],
            "properties": {
                "branch": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
    }
}